+++ src/Atoms/CssGrid/CssGrid.tsx
@@ -12,7 +12,7 @@ import {
   Size,
 } from './CssGrid.types';
 import { isUndefined, assert } from '../../common/utils';
-import { getAreasInfo } from './utils';
+import { getAreasInfo, getMsRawTemplateColumnOrRowStyles } from './utils';
 
 const formatAreas = (areas: Props['areas']) =>
   areas
@@ -161,7 +161,7 @@ const generateMSGridStyles = ({
        * space-separated set of css units
        * @todo verify this hypothesis
        */
-      const msGridRowsStyles = rawTemplateRowsStyles.split(' ').join(` ${rowGutter} `);
+      const msGridRowsStyles = getMsRawTemplateColumnOrRowStyles(rawTemplateRowsStyles, rowGutter);
       const msGridRows = `-ms-grid-rows: ${msGridRowsStyles};`;
       styles.push(msGridRows);
     } else {
@@ -179,7 +179,10 @@ const generateMSGridStyles = ({
        * space-separated set of css units
        * @todo verify this hypothesis
        */
-      const msGridColumnsStyles = rawTemplateColumnStyles.split(' ').join(` ${colGutter} `);
+      const msGridColumnsStyles = getMsRawTemplateColumnOrRowStyles(
+        rawTemplateColumnStyles,
+        colGutter,
+      );
       const msGridColumns = `-ms-grid-columns: ${msGridColumnsStyles};`;
       styles.push(msGridColumns);
     } else {