+++ src/Molecules/Input/Checkbox/Checkbox.tsx
@@ -26,6 +26,9 @@ const CheckmarkBox = styled(Flexbox)`
     width: 100%;
     height: 100%;
     position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
   }
 `;