+++ src/theme/theme.types.ts
@@ -149,6 +149,8 @@ export type Theme = {
     orderDepthBackground: RawColor['gray6'];
     /** gray5 */
     orderDepthDarkBackground: RawColor['gray5'];
+    /** complementaryBlue2 */
+    loanRatesGraphColor2: RawColor['complementaryBlue2'];
   };
 
   spacing: {